Rename google-pubsub Filebeat input to gcp-pubsub
