+++ x-pack/filebeat/module/googlecloud/audit/config/input.yml
@@ -1,6 +1,6 @@
-{{ if eq .input "google-pubsub" }}
+{{ if eq .input "gcp-pubsub" }}
 
-type: google-pubsub
+type: gcp-pubsub
 project_id: {{ .project_id }}
 topic: {{ .topic }}
 subscription.name: {{ .subscription_name }}
