+++ x-pack/filebeat/module/googlecloud/audit/manifest.yml
@@ -2,7 +2,7 @@ module_version: "1.0"
 
 var:
   - name: input
-    default: google-pubsub
+    default: gcp-pubsub
   - name: project_id
     default: SET_PROJECT_NAME
   - name: topic
